moves all the monitor logic into the einstein api
