+++ einstein_api/static/js/einstein_api/controllers/pair_monitor.js
