+++ somno/__init__.py
@@ -15,8 +15,6 @@ class Application(application.OpalApplication):
         'js/somno/controllers/newgraph.js',
         'js/somno/controllers/induction_drug_controller.js',
         'js/somno/controllers/new_infusions.js',
-        'js/somno/controllers/pair_monitor.js',
-        'js/somno/controllers/unpair_monitor.js',
     ]
     styles = [
         "css/somno.css",
