+++ somno/metadata.py
@@ -1,6 +1,4 @@
-from django.conf import settings
 from opal.core import metadata
-from somno.models import Monitor
 
 
 class DrugTypes(metadata.Metadata):
@@ -45,16 +43,3 @@ class DrugTypes(metadata.Metadata):
                     ],
             }
         }
-
-
-class Monitors(metadata.Metadata):
-    slug = 'monitors'
-
-    @classmethod
-    def to_dict(klass, *args, **kwargs):
-        id_user_machine_name = Monitor.objects.values_list(
-            "id", "user_machine_name"
-        )
-        return dict(monitors={
-            i: v for i, v in id_user_machine_name
-        })
