+++ somno/models.py
@@ -164,20 +164,3 @@ class AnaestheticAssesment(models.EpisodeSubrecord):
     General_Risks = db_models.TextField(blank=True, null=True,)
     AdditionalRisks = db_models.TextField(blank=True, null=True)
     TimeSeen = db_models.DateTimeField(blank=True, null=True,)
-
-
-class Monitor(db_models.Model):
-    user_machine_name = db_models.CharField(max_length=256, unique=True)
-    einstein_id = db_models.CharField(max_length=256, unique=True)
-
-    def __str__(self):
-        return "{} - {}".format(self.user_machine_name, self.einstein_id)
-
-
-class MonitorPatientPairing(models.PatientSubrecord):
-    start = db_models.DateTimeField(blank=True, null=True)
-    stop = db_models.DateTimeField(blank=True, null=True)
-    monitor = db_models.ForeignKey(Monitor)
-
-    def monitor_options(self):
-        return Monitor.objects.all()
